# openapi.yaml
openapi: 3.0.0
info:
  title: Les Chauffagistes mining pool API
  version: 1.0.0
  description: API Les Chauffagistes pour accéder aux métriques de minage
...
paths:
...
  /v1/stats/{user}:
    parameters:
      - name: user
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getUserStats
      tags: [pool]
      summary: User stats
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "./vendor/chauff-cmn.yaml#/components/schemas/PoolStats"
        "404":
          description: Adresse introuvable
          content:
            application/json:
              schema:
                type: object
                required: [error]
                properties:
                  error:
                    type: string
                    example: Adresse introuvable
...